+++ ftplugin/java.lua
@@ -13,13 +13,13 @@ local extendedClientCapabilities = jdtls.extendedClientCapabilities
 
 local bundles = {
   vim.fn.glob(
-    home
-      .. "/.local/share/nvim/mason/packages/java-debug/com.microsoft.java.debug.plugin/target/com.microsoft.java.debug.plugin-*.jar"
+    home .. "/.local/share/nvim/mason/packages/java-debug-adapter/extension/server/com.microsoft.java.debug.plugin-*.jar",
+    true
   ),
 }
 vim.list_extend(
   bundles,
-  vim.split(vim.fn.glob(home .. "/.local/share/nvim/mason/packages/vscode-java-test/server/*.jar", true), "\n")
+  vim.split(vim.fn.glob(home .. "/.local/share/nvim/mason/packages/java-test/extension/server/*.jar", true), "\n")
 )
 
 local config = {